+++ CHANGELOG.md
@@ -59,6 +59,7 @@
 39. Added Launch Option "--melonloader.agfvunity" to Force the Version of Unity Dependencies to use.
 40. Fixed Assembly Resolve Issue when a Plugin references a Mod.
 41. Fixed Displacement Issue when calling Save for MelonPreferences or Legacy MelonPrefs.
+42. Fixed Issue with Legacy MelonPrefs logging "Legacy Config Saved!" even when nothing happened.
 
 ---